+++ kurly-core/src/main/java/com/devcourse/kurlymurly/application/product/ProductFacade.java
@@ -5,11 +5,11 @@ import com.devcourse.kurlymurly.domain.product.Product;
 import com.devcourse.kurlymurly.domain.product.ProductDomain;
 import com.devcourse.kurlymurly.domain.product.SupportDomain;
 import com.devcourse.kurlymurly.domain.product.review.Review;
+import com.devcourse.kurlymurly.domain.service.OrderService;
 import com.devcourse.kurlymurly.domain.service.ProductCommand;
 import com.devcourse.kurlymurly.domain.service.ProductQuery;
 import com.devcourse.kurlymurly.domain.service.ReviewCommand;
 import com.devcourse.kurlymurly.domain.service.ReviewQuery;
-import com.devcourse.kurlymurly.module.order.service.OrderService;
 import com.devcourse.kurlymurly.web.product.FavoriteResponse;
 import com.devcourse.kurlymurly.web.product.ProductRequest;
 import com.devcourse.kurlymurly.web.product.ProductResponse;
