--- kurly-domain/build.gradle
@@ -1,10 +1,0 @@
-dependencies {
-    api project(":kurly-domain:kurly-db")
-
-    implementation project(":kurly-domain:kurly-user")
-    implementation project(':kurly-support:kurly-common-support')
-    implementation project(':kurly-support:kurly-web-support')
-}
-
-bootJar.enabled=false
-jar.enabled=true
