+++ kurly-domain/kurly-order/src/main/java/com/devcourse/kurlymurly/domain/order/state/OrderState.java
@@ -1,0 +1,6 @@
+package com.devcourse.kurlymurly.domain.order.state;
+
+public interface OrderState {
+    OrderState nextState();
+    OrderState cancel();
+}
