+++ kurly-domain/kurly-order/src/main/java/com/devcourse/kurlymurly/domain/order/state/Ordered.java
@@ -1,0 +1,13 @@
+package com.devcourse.kurlymurly.domain.order.state;
+
+public class Ordered implements OrderState {
+    @Override
+    public OrderState nextState() {
+        return new Processing();
+    }
+
+    @Override
+    public OrderState cancel() {
+        return new Canceled();
+    }
+}
