+++ kurly-domain/kurly-order/src/test/java/com/devcourse/kurlymurly/domain/UserFixture.java
@@ -1,0 +1,34 @@
+package com.devcourse.kurlymurly.domain;
+
+import com.devcourse.kurlymurly.domain.user.User;
+import com.devcourse.kurlymurly.domain.user.UserInfo;
+
+import java.time.LocalDate;
+
+public enum UserFixture {
+    USER_FIXTURE("수연장", "abd1234", "kurly1234", "aaaa@gmail.com",
+            new UserInfo(LocalDate.now(), "문희조", "MAN"), "1234-4567");
+
+    private final String name;
+    private final String loginId;
+    private final String password;
+    private final String email;
+    private final UserInfo userInfo;
+    private final String phoneNumber;
+
+    UserFixture(String name, String loginId, String password, String email, UserInfo userInfo, String phoneNumber) {
+        this.name = name;
+        this.loginId = loginId;
+        this.password = password;
+        this.email = email;
+        this.userInfo = userInfo;
+        this.phoneNumber = phoneNumber;
+    }
+
+    public User toEntity() {
+        User user =  new User(name, loginId, password, email, userInfo, phoneNumber);
+        user.updatePayPassword("123456");
+
+        return user;
+    }
+}
