--- kurly-domain/src/main/java/com/devcourse/kurlymurly/module/order/domain/state/OrderState.java
@@ -1,6 +1,0 @@
-package com.devcourse.kurlymurly.module.order.domain.state;
-
-public interface OrderState {
-    OrderState nextState();
-    OrderState cancel();
-}
