--- kurly-domain/src/main/java/com/devcourse/kurlymurly/module/order/domain/state/Ordered.java
@@ -1,13 +1,0 @@
-package com.devcourse.kurlymurly.module.order.domain.state;
-
-public class Ordered implements OrderState {
-    @Override
-    public OrderState nextState() {
-        return new Processing();
-    }
-
-    @Override
-    public OrderState cancel() {
-        return new Canceled();
-    }
-}
