--- kurly-domain/src/main/java/com/devcourse/kurlymurly/module/order/domain/state/Processing.java
@@ -1,14 +1,0 @@
-package com.devcourse.kurlymurly.module.order.domain.state;
-
-public class Processing implements OrderState {
-    @Override
-    public OrderState nextState() {
-        return new Delivering();
-    }
-
-    @Override
-    public OrderState cancel() {
-        // TODO: 리뷰 작성 제한,, 어떻게,,
-        return new Canceled();
-    }
-}
