--- kurly-domain/src/main/java/com/devcourse/kurlymurly/module/order/domain/support/OrderSupportRepository.java
@@ -1,10 +1,0 @@
-package com.devcourse.kurlymurly.module.order.domain.support;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface OrderSupportRepository extends JpaRepository<OrderSupport, Long> {
-    List<OrderSupport> findAllByUserId(Long userId);
-    List<OrderSupport> findByOrderNumber(String orderNumber);
-}
